Bounds for the KL-divergence

Computation of the bounds of the KL-divergence for variations of each parameter of a CI object.

Usage

KL_bounds(ci, delta)

Arguments

ci

object of class CI.

delta

multiplicative variation coefficient for the entry of the covariance matrix given in entry.

Value

A dataframe including the KL-divergence bound for each co-variation scheme (model-preserving and standard) and every entry of the covariance matrix. For variations leading to non-positive semidefinite matrix, the dataframe includes a NA.

Details

Let \(\Sigma\) be the covariance matrix of a Gaussian Bayesian network with \(n\) vertices. Let \(D\) and \(\Delta\) be variation matrices acting additively on \(\Sigma\). Let also \(\tilde\Delta\) be a model-preserving co-variation matrix. Denote with \(Y\) and \(\tilde{Y}\) the original and the perturbed random vectors. Then for a standard sensitivity analysis $$KL(\tilde{Y}||Y)\leq 0.5n\max\left\{f(\lambda_{\max}(D\Sigma^{-1})),f(\lambda_{\min}(D\Sigma^{-1}))\right\}$$ whilst for a model-preserving one $$KL(\tilde{Y}||Y)\leq 0.5n\max\left\{f(\lambda_{\max}(\tilde\Delta\circ\Delta)),f(\lambda_{\min}(\tilde\Delta\circ\Delta))\right\}$$ where \(\lambda_{\max}\) and \(\lambda_{\min}\) are the largest and the smallest eigenvalues, respectively, \(f(x)=\ln(1+x)-x/(1+x)\) and \(\circ\) denotes the Schur or element-wise product.

References

C. Görgen & M. Leonelli (2020), Model-preserving sensitivity analysis for families of Gaussian distributions. Journal of Machine Learning Research, 21: 1-32.

See also

Examples

KL_bounds(synthetic_ci,1.05)
#>    row col   standard    total  partial row_based col_based
#> 1    1   1 0.06319007 1.681933 1.622910  1.622910  1.622910
#> 2    1   2 0.71822823 1.681933 1.646925  1.630979  1.630979
#> 3    1   3 0.26097348 1.681933 1.646925  1.634914  1.634914
#> 4    1   4         NA 1.681933       NA        NA        NA
#> 5    2   2 0.26208872 1.681933 1.646925  1.630979  1.630979
#> 6    2   3 2.56494598 1.681933 1.646925  1.634914  1.634914
#> 7    2   4         NA 1.681933       NA        NA        NA
#> 8    3   3 0.63258146 1.681933 1.622910  1.622910  1.622910
#> 9    3   4         NA 1.681933       NA        NA        NA
#> 10   4   4 1.32814438 1.681933 1.622910  1.622910  1.622910