bnmonitor: A package for sensitivity analysis and robustness in Bayesian networks

Sensitivity and robustness analysis for Bayesian networks.

Details

bnmonitor provides functions to perform sensitivity analysis for both discrete Bayesian networks (DBNs) and Gaussian Bayesian networks (GBNs).

In the discrete case, it provides three categories of functions: co-variation schemes, dissimilarity measures and sensitivity related functions.

In the continuous case, both standard and model-preserving methods are available for perturbation of the mean vector and the co-variance matrix.

bnmonitor further provides function to perform robustness studies in DBNs to verify how well a network fits a specific dataset.

DBNs - Robustness

The available functions for robustness are:

Node monitors (node_monitor): contribution of each vertex to the overall log-likelihood of the model.

Observation's influence (influential_obs): difference in the log-likelihood of a model learnt with the full dataset and with all but one observation.

Final node monitors (node_monitor): marginal and conditional node monitors to assess the fit of a vertex distribution to the full dataset.

Sequential node monitors (seq_node_monitor): marginal and conditional node monitors for a specific vertex only using sequentially subsets of the dataset.

Sequential parent-child monitor (seq_pa_ch_monitor): parent-child node monitor for a specific vertex and a specific configuration of its parents using sequentially subsets of the dataset.

DBNs - Co-variation schemes

The available co-variation schemes are:

Uniform co-variation scheme (uniform_covar): distributes the probability mass to be co-varied uniformly among the co-varying parameters.

Proportional co-variation scheme (proportional_covar): distributes the probability mass to be co-varied in the same proportion as in the original Bayesian network.

Order-preserving co-variation scheme (orderp_covar):distributes the to be co-varied probability mass among the co-varying parameters so that the original order of parameters is preserved.

DBNs - Dissimilarity measures

The dissimilarity measures quantify the difference between a Bayesian network and its update after parameter variation.
The available dissimilarity measures are:

Chan-Darwiche distance (CD)

Kullback-Leibler divergence (KL)

DBNs - Sensitivity functions

The available functions for sensitivity analysis are:

Sensitivity function (sensitivity): returns a certain probability of interest given a parameter change. Evidence can be considered.

Sensitivity query (sensquery): returns the parameter changes needed to get a certain probability of interest. Evidence can be considered.

GBNs - Model-Preserving matrices

The available functions to construct model-preserving co-variation matrices are:

Total co-variation matrix (total_covar_matrix).

Partial co-variation matrix (partial_covar_matrix).

Row-based co-variation matrix (row_covar_matrix).

Column-based co-variation matrix (col_covar_matrix).

GBNs - Mean and Covariance variations

The available functions to perturb the distribution of a GBN are:

Mean variations (mean_var).

Standard covariance variations (covariance_var).

Model-preserving covariance variations (model_pres_cov).

GBNs - Dissimilarity measures

The available dissimilarity measures are:

Frobenius norm (Fro).

Jeffrey's distance (Jeffreys).

Kullback-Leibler divergence (KL).

Upper bound to the KL divergence (KL_bounds).